consolidate multi
