+++ src/stream_ml/core/multi/__init__.py
@@ -1,0 +1,3 @@
+"""Multiple models."""
+
+__all__: list[str] = []
