+++ src/stream_ml/core/multi/bases.py
